libdnf-cli/output/repo_info.hpp: Cast the format_size argument to signed
Even though the argument type of format_size should perhaps be unsigned,
it triggers a cascade of warnings, proliferating through the progressbar
widget classes. Might need a review of the type used there.

+++ libdnf-cli/output/repo_info.hpp
@@ -107,7 +107,7 @@ void RepoInfo::add_repo(const Repo & repo, bool verbose, bool show_sack_data) {
         add_line("Available packages", repo.get_available_pkgs(), nullptr, group_repodata);
         add_line("Total packages", repo.get_pkgs(), nullptr, group_repodata);
 
-        std::string size = libdnf::cli::utils::units::format_size(repo.get_size());
+        std::string size = libdnf::cli::utils::units::format_size(static_cast<int64_t>(repo.get_size()));
         add_line("Size", size, nullptr, group_repodata);
 
         add_line("Content tags", repo.get_content_tags(), nullptr, group_repodata);
